refactor(delegate): move queue + flush RPCs to authenticated main Service
