+++ api-spec/openapi/swagger/fulmine/v1/service.swagger.json
@@ -169,6 +169,62 @@
         ]
       }
     },
+    "/v1/delegate/queue": {
+      "get": {
+        "summary": "GetDelegateQueue returns intents currently held for coalesced registration",
+        "operationId": "Service_GetDelegateQueue",
+        "responses": {
+          "200": {
+            "description": "A successful response.",
+            "schema": {
+              "$ref": "#/definitions/v1GetDelegateQueueResponse"
+            }
+          },
+          "default": {
+            "description": "An unexpected error response.",
+            "schema": {
+              "$ref": "#/definitions/rpcStatus"
+            }
+          }
+        },
+        "tags": [
+          "Service"
+        ]
+      }
+    },
+    "/v1/delegate/queue/flush": {
+      "post": {
+        "summary": "FlushDelegateQueue registers all currently-held intents immediately",
+        "operationId": "Service_FlushDelegateQueue",
+        "responses": {
+          "200": {
+            "description": "A successful response.",
+            "schema": {
+              "$ref": "#/definitions/v1FlushDelegateQueueResponse"
+            }
+          },
+          "default": {
+            "description": "An unexpected error response.",
+            "schema": {
+              "$ref": "#/definitions/rpcStatus"
+            }
+          }
+        },
+        "parameters": [
+          {
+            "name": "body",
+            "in": "body",
+            "required": true,
+            "schema": {
+              "$ref": "#/definitions/v1FlushDelegateQueueRequest"
+            }
+          }
+        ],
+        "tags": [
+          "Service"
+        ]
+      }
+    },
     "/v1/delegates": {
       "get": {
         "summary": "ListDelegates returns delegate tasks filtered by status, paginated by limit/offset.",
@@ -1194,6 +1250,20 @@
         }
       }
     },
+    "v1DelegateQueueEntry": {
+      "type": "object",
+      "properties": {
+        "taskId": {
+          "type": "string"
+        },
+        "registerBy": {
+          "type": "string",
+          "format": "int64",
+          "title": "Unix seconds by which this intent must be registered"
+        }
+      },
+      "title": "DelegateQueueEntry describes a single intent buffered for coalesced registration"
+    },
     "v1DelegateRefundParams": {
       "type": "object",
       "properties": {
@@ -1207,6 +1277,14 @@
           "type": "string"
         }
       }
+    },
+    "v1FlushDelegateQueueRequest": {
+      "type": "object",
+      "title": "FlushDelegateQueueRequest is the request message for FlushDelegateQueue rpc"
+    },
+    "v1FlushDelegateQueueResponse": {
+      "type": "object",
+      "title": "FlushDelegateQueueResponse is the response message for FlushDelegateQueue rpc"
     },
     "v1GetAddressResponse": {
       "type": "object",
@@ -1227,6 +1305,29 @@
           "format": "uint64"
         }
       }
+    },
+    "v1GetDelegateQueueResponse": {
+      "type": "object",
+      "properties": {
+        "entries": {
+          "type": "array",
+          "items": {
+            "type": "object",
+            "$ref": "#/definitions/v1DelegateQueueEntry"
+          }
+        },
+        "nextFlushAt": {
+          "type": "string",
+          "format": "int64",
+          "title": "Unix seconds of the next scheduled flush; 0 if the queue is empty"
+        },
+        "coalesceWindow": {
+          "type": "string",
+          "format": "int64",
+          "title": "Configured coalescing window in seconds"
+        }
+      },
+      "title": "GetDelegateQueueResponse is the response message for GetDelegateQueue rpc"
     },
     "v1GetInfoResponse": {
       "type": "object",
