+++ internal/interface/grpc/handlers/service_handler.go
@@ -20,11 +20,12 @@ import (
 )
 
 type serviceHandler struct {
-	svc *application.Service
+	svc         *application.Service
+	delegateSvc *application.DelegateService
 }
 
-func NewServiceHandler(svc *application.Service) pb.ServiceServer {
-	return &serviceHandler{svc}
+func NewServiceHandler(svc *application.Service, delegateSvc *application.DelegateService) pb.ServiceServer {
+	return &serviceHandler{svc, delegateSvc}
 }
 
 func (h *serviceHandler) GetAddress(
@@ -729,6 +730,38 @@ func chainSwapStatusToString(status domain.ChainSwapStatus) string {
 	default:
 		return "unknown"
 	}
+}
+
+func (h *serviceHandler) GetDelegateQueue(
+	ctx context.Context, req *pb.GetDelegateQueueRequest,
+) (*pb.GetDelegateQueueResponse, error) {
+	if h.delegateSvc == nil {
+		return nil, status.Error(codes.FailedPrecondition, "delegation not enabled")
+	}
+	entries, nextFlush := h.delegateSvc.GetQueue()
+	pbEntries := make([]*pb.DelegateQueueEntry, len(entries))
+	for i, e := range entries {
+		pbEntries[i] = &pb.DelegateQueueEntry{TaskId: e.TaskID, RegisterBy: e.RegisterBy.Unix()}
+	}
+	var nextFlushAt int64
+	if !nextFlush.IsZero() {
+		nextFlushAt = nextFlush.Unix()
+	}
+	return &pb.GetDelegateQueueResponse{
+		Entries:        pbEntries,
+		NextFlushAt:    nextFlushAt,
+		CoalesceWindow: int64(h.delegateSvc.CoalesceWindow().Seconds()),
+	}, nil
+}
+
+func (h *serviceHandler) FlushDelegateQueue(
+	ctx context.Context, req *pb.FlushDelegateQueueRequest,
+) (*pb.FlushDelegateQueueResponse, error) {
+	if h.delegateSvc == nil {
+		return nil, status.Error(codes.FailedPrecondition, "delegation not enabled")
+	}
+	h.delegateSvc.FlushRegistrationQueue()
+	return &pb.FlushDelegateQueueResponse{}, nil
 }
 
 func (h *serviceHandler) ListDelegates(
